Add new routing classes for multithreading and multiprocessing
This commit adds the IMessageRouter, IMessageSender and IMessageReceiver
interfaces, and implementations of those interfaces intended for
background tasks running (a) in a background thread, or (b) in a
separate child process.

The new classes are similar in role to the old ones. A key
difference is that the message router is no longer responsible for
determining when to dispose of a receiver; instead, the code
that requests the (sender, receiver) pair is determining when
the receiver is no longer needed.