Transition executor machinery to new routers
This commit updates the TraitsExecutor, MultithreadingContext and
FutureWrapper to use the new message routers instead of the old
ones.

Notable change: we introduce a new "done" trait to the `FutureWrapper`
to allow us to listen directly to the set of wrappers in order
to determine when a future is complete. This is necessary so that
we can ask the message router to unregister the receiver.
(Previously, the router took care of figuring out when to unregister
receivers, but the new router moves that responsibility to the
code that requests the receiver.)

The TraitsExecutor._wrappers trait is now simply a set of
`FutureWrapper` objects; we no longer need direct access to the
futures themselves in the executor.