DEV : Replace on_trait_change decorator with observe
This commit replaces the on_trait_change decorator with the observe
decorator. The changes have been made in the documentation, the
testsuite and the package itself.

Note that the new decorator requires a change to the api of the
decorated method - which now needs to accept a single "event" argument
instead of the earlier signature.